+++ R/findVariableLists.R
@@ -36,7 +36,8 @@ findVariableLists <- function(jdx, debug = 0) {
     "XII", # imaginary NMR
     "NMR_2D", # real 2D NMR in NTUPLE format
     "LC_MS", # LC or GC-MS data in NTUPLE format
-    "PEAK_TABLE" # Simple peak list; AFFN assumed
+    "XYXY", # PEAK TABLE, equal to the next one in practice
+    "XYXY" # XYXY
   )
 
   nf <- length(VL_fmts)
@@ -49,7 +50,8 @@ findVariableLists <- function(jdx, debug = 0) {
     "^\\s*##PAGE\\s*=\\s*N=2",
     "^\\s*##PAGE\\s*=\\s*F1=",
     "^\\s*##PAGE\\s*=\\s*T=",
-    "^\\s*##PEAK TABLE\\s*=\\s*\\(XY\\.\\.XY\\)"
+    "^\\s*##PEAK TABLE\\s*=\\s*\\(XY\\.\\.XY\\)",
+    "^\\s*##XYPOINTS\\s*=\\s*\\(XY\\.\\.XY\\)"
   )
 
   # Add other END patterns here (each associated with a specific VL_fmts entry)
@@ -59,8 +61,13 @@ findVariableLists <- function(jdx, debug = 0) {
     "^\\s*##END\\s{1}NTUPLES\\s*=",
     "^\\s*##PAGE\\s*=\\s*F1=", # In NTUPLES there are several/many of these
     "^\\s*##PAGE\\s*=\\s*T=", # In NTUPLES there are several/many of these
+    "^\\s*##END\\s*=",
     "^\\s*##END\\s*="
   )
+
+  # Developer sanity checks:
+  if (length(ST_pats) != nf) stop("Wrong number of ST_pats")
+  if (length(END_pats) != nf) stop("Wrong number of END_pats")
 
   # Find the beginning & end of each variable list.
   # We are checking for any and all formats in the file
@@ -111,7 +118,7 @@ findVariableLists <- function(jdx, debug = 0) {
 
   # Find all comment-only lines exclusive of metadata; these cause a variety of problems.
   # Keep original line numbers. CURRENTLY NOT USED OTHER THAN THIS FUNCTION, but is returned.
-  # There is a multi-line comment label "##=" but we do not search for this (not sure I've seen it).
+  # There is a multi-line comment label "##=" but we do not search for this (not sure I've ever seen it).
 
   comOnly <- grep("^\\$\\$", jdx)
   comOnly <- setdiff(comOnly, 1:(spec_st[1] - 1))
@@ -172,10 +179,16 @@ findVariableLists <- function(jdx, debug = 0) {
   VL[[3]] <- comOnly
 
   for (i in 4:length(VL)) {
-    VL[[i]] <- c(DF$Format[i - 2], jdx[DF$FirstLine[i - 2]:DF$LastLine[i - 2]])
+    st_line <- DF$FirstLine[i - 2]
+    end_line <- DF$LastLine[i - 2]
+    keep_lines <- st_line:end_line
+    # first entry is fmt (for later dispatch to the right processing)
+    VL[[i]] <- c(DF$Format[i - 2], jdx[keep_lines])
+    name_vec <- c("fmt", paste("Line", keep_lines, sep = "_"))
+    names(VL[[i]]) <-name_vec # name it for debugging purposes downstream
   }
 
-  # The generic VL_1 names are replaced when these results are passed back to readJDX
+  # The generic VL_X names are replaced when these results are passed back to readJDX
   names(VL) <- c("DataGuide", "Metadata", "Comments", paste("VL", 1:(length(VL) - 3), sep = "_"))
   return(VL)
 }