+++ inst/doc/GuideToVignettes.R
@@ -3,3 +3,6 @@ library("readJDX")
 rm(list = ls())
 desc <- packageDescription("readJDX")
 
+## ---- eval = FALSE------------------------------------------------------------
+#  browseVignettes("readJDX")
+